spi: cadence-quadspi: Handle multiple clocks
The QSPI hardware in some SoCs (such as the StarFive JH7100) requires
multiple clocks to function. We handle this case by enabling all clocks
assigned to the device, and find the reference clock by the name
"refclk".

As backwards compatibility, if only one clock is assigned to the device,
we assume that it is the refclk.

Signed-off-by: Vivian Wang <dramforever@live.com>
Co-Developed-by: Icenowy Zheng <uwu@icenowy.me>
Signed-off-by: Icenowy Zheng <uwu@icenowy.me>